+++ frontend/assets/svgs/index.ts
@@ -1,0 +1,6 @@
+import ArrowSVG from './arrow.svg';
+import CloseSVG from './close.svg';
+import MagnifyingGlassSVG from './magnifyingGlass.svg';
+import WhiteCheckSVG from './whiteCheck.svg';
+
+export { ArrowSVG, CloseSVG, MagnifyingGlassSVG, WhiteCheckSVG };
