+++ frontend/components/common/checkbox/index.tsx
@@ -3,7 +3,7 @@
 import Image from 'next/image';
 import { ChangeEventHandler, useState } from 'react';
 
-import WhiteCheckSVG from '@/assets/svgs/white_check.svg';
+import { WhiteCheckSVG } from '@/assets/svgs';
 
 import { Variant } from '@/types/components';
