+++ frontend/components/filter/jobsFilter/components/categories.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 
-import ArrowSVG from '@/assets/svgs/arrow.svg';
+import { ArrowSVG } from '@/assets/svgs';
 
 import { Job } from '@/types/api/jobs';
